Add explicit activation replacement in all mlp blocks of transformers
If some versions of transformers, a single instance of the activation
function is shared among all layers, so it appears in the named_modules
a single time. Changing it with the recursive approach results in
replacting only the activation at the first layer. Explicit replacement
fixes this issue, but depends on knowledge of model inner structure.